+++ ChangeLog.md
@@ -3,7 +3,8 @@
 ## Unreleased
 
 - Fix a bug that caused another code-leftover to be displayed in the finalization proof
-  line of the gas rewards section of the `consensus show-chain-parameters` output.
+  line of the gas rewards section of the `consensus show-chain-parameters` output. This
+  is only when the queried chain runs protocol version 6 or lower.
 
 ## 5.1.1